create parent Person class and file w/ initialize, high fives, make sure rspecs still pass
