+++ spec/apprentice_teacher_spec.rb
@@ -1,3 +1,4 @@
+require_relative '../person.rb'
 require_relative '../apprentice_teacher'
 
 describe ApprenticeTeacher do
