+++ spec/student_spec.rb
@@ -1,3 +1,4 @@
+require_relative '../person.rb'
 require_relative '../student'
 
 describe Student do
